cloud_functions/rag_ingestion/requirements.txt
# Cloud Function dependencies
functions-framework==3.*
google-cloud-storage==2.*
google-cloud-aiplatform>=1.46.0,<2.0.0
vertexai>=1.46.0,<2.0.0
google-cloud-logging==3.*